requirements.txt
lxml~=4.8.0
aiohttp~=3.7.4.post0
requests~=2.25.1
urllib3~=1.25.8
matplotlib~=3.5.1
pytz~=2022.1
pycountry~=22.3.5
pandas~=1.4.1
openpyxl~=3.0.9
redis~=3.5.3
pymongo~=3.11.2
motor~=2.4.0
dnspython~=2.1.0
aiosqlite~=0.17.0
influxdb-client~=1.17.0
pymysql~=1.0.2
Pillow~=9.0.1
CairoSVG~=2.5.2
tabulate~=0.8.9
discordspy~=0.1.2
aiogram~=2.15